+++ packages/@aws-cdk/aws-cognito/lib/index.ts
@@ -5,4 +5,3 @@ export * from './user-pool-attr';
 export * from './user-pool-client';
 export * from './user-pool-domain';
 export * from './user-pool-idp';
-export * from './user-pool-idps';
