--- packages/@aws-cdk/aws-cognito/lib/user-pool-idps/index.ts
@@ -1,4 +1,0 @@
-export * from './base';
-export * from './amazon';
-export * from './facebook';
-export * from './google';
