+++ packages/@aws-cdk/aws-cognito/test/user-pool-idp.test.ts
@@ -1,0 +1,295 @@
+import '@aws-cdk/assert/jest';
+import { Stack } from '@aws-cdk/core';
+import { ProviderAttribute, UserPool, UserPoolIdentityProviderAmazon, UserPoolIdentityProviderFacebook, UserPoolIdentityProviderGoogle } from '../lib';
+
+describe('UserPoolIdentityProvider', () => {
+  describe('amazon', () => {
+    test('defaults', () => {
+      // GIVEN
+      const stack = new Stack();
+      const pool = new UserPool(stack, 'userpool');
+
+      // WHEN
+      new UserPoolIdentityProviderAmazon(stack, 'userpoolidp', {
+        userPool: pool,
+        clientId: 'amzn-client-id',
+        clientSecret: 'amzn-client-secret',
+      });
+
+      expect(stack).toHaveResource('AWS::Cognito::UserPoolIdentityProvider', {
+        ProviderName: 'LoginWithAmazon',
+        ProviderType: 'LoginWithAmazon',
+        ProviderDetails: {
+          client_id: 'amzn-client-id',
+          client_secret: 'amzn-client-secret',
+          authorize_scopes: 'profile',
+        },
+      });
+    });
+
+    test('scopes', () => {
+      // GIVEN
+      const stack = new Stack();
+      const pool = new UserPool(stack, 'userpool');
+
+      // WHEN
+      new UserPoolIdentityProviderAmazon(stack, 'userpoolidp', {
+        userPool: pool,
+        clientId: 'amzn-client-id',
+        clientSecret: 'amzn-client-secret',
+        scopes: ['scope1', 'scope2'],
+      });
+
+      expect(stack).toHaveResource('AWS::Cognito::UserPoolIdentityProvider', {
+        ProviderName: 'LoginWithAmazon',
+        ProviderType: 'LoginWithAmazon',
+        ProviderDetails: {
+          client_id: 'amzn-client-id',
+          client_secret: 'amzn-client-secret',
+          authorize_scopes: 'scope1 scope2',
+        },
+      });
+    });
+
+    test('registered with user pool', () => {
+      // GIVEN
+      const stack = new Stack();
+      const pool = new UserPool(stack, 'userpool');
+
+      // WHEN
+      const provider = new UserPoolIdentityProviderAmazon(stack, 'userpoolidp', {
+        userPool: pool,
+        clientId: 'amzn-client-id',
+        clientSecret: 'amzn-client-secret',
+      });
+
+      // THEN
+      expect(pool.identityProviders).toContain(provider);
+    });
+
+    test('attribute mapping', () => {
+      // GIVEN
+      const stack = new Stack();
+      const pool = new UserPool(stack, 'userpool');
+
+      // WHEN
+      new UserPoolIdentityProviderAmazon(stack, 'userpoolidp', {
+        userPool: pool,
+        clientId: 'amazn-client-id',
+        clientSecret: 'amzn-client-secret',
+        attributeMapping: {
+          givenName: ProviderAttribute.AMAZON_NAME,
+          address: ProviderAttribute.other('amzn-address'),
+          custom: {
+            customAttr1: ProviderAttribute.AMAZON_EMAIL,
+            customAttr2: ProviderAttribute.other('amzn-custom-attr'),
+          },
+        },
+      });
+
+      // THEN
+      expect(stack).toHaveResource('AWS::Cognito::UserPoolIdentityProvider', {
+        AttributeMapping: {
+          given_name: 'name',
+          address: 'amzn-address',
+          customAttr1: 'email',
+          customAttr2: 'amzn-custom-attr',
+        },
+      });
+    });
+  });
+
+  describe('facebook', () => {
+    test('defaults', () => {
+      // GIVEN
+      const stack = new Stack();
+      const pool = new UserPool(stack, 'userpool');
+
+      // WHEN
+      new UserPoolIdentityProviderFacebook(stack, 'userpoolidp', {
+        userPool: pool,
+        clientId: 'fb-client-id',
+        clientSecret: 'fb-client-secret',
+      });
+
+      expect(stack).toHaveResource('AWS::Cognito::UserPoolIdentityProvider', {
+        ProviderName: 'Facebook',
+        ProviderType: 'Facebook',
+        ProviderDetails: {
+          client_id: 'fb-client-id',
+          client_secret: 'fb-client-secret',
+          authorize_scopes: 'public_profile',
+        },
+      });
+    });
+
+    test('scopes & api_version', () => {
+      // GIVEN
+      const stack = new Stack();
+      const pool = new UserPool(stack, 'userpool');
+
+      // WHEN
+      new UserPoolIdentityProviderFacebook(stack, 'userpoolidp', {
+        userPool: pool,
+        clientId: 'fb-client-id',
+        clientSecret: 'fb-client-secret',
+        scopes: ['scope1', 'scope2'],
+        apiVersion: 'version1',
+      });
+
+      expect(stack).toHaveResource('AWS::Cognito::UserPoolIdentityProvider', {
+        ProviderName: 'Facebook',
+        ProviderType: 'Facebook',
+        ProviderDetails: {
+          client_id: 'fb-client-id',
+          client_secret: 'fb-client-secret',
+          authorize_scopes: 'scope1,scope2',
+          api_version: 'version1',
+        },
+      });
+    });
+
+    test('registered with user pool', () => {
+      // GIVEN
+      const stack = new Stack();
+      const pool = new UserPool(stack, 'userpool');
+
+      // WHEN
+      const provider = new UserPoolIdentityProviderFacebook(stack, 'userpoolidp', {
+        userPool: pool,
+        clientId: 'fb-client-id',
+        clientSecret: 'fb-client-secret',
+      });
+
+      // THEN
+      expect(pool.identityProviders).toContain(provider);
+    });
+
+    test('attribute mapping', () => {
+      // GIVEN
+      const stack = new Stack();
+      const pool = new UserPool(stack, 'userpool');
+
+      // WHEN
+      new UserPoolIdentityProviderFacebook(stack, 'userpoolidp', {
+        userPool: pool,
+        clientId: 'fb-client-id',
+        clientSecret: 'fb-client-secret',
+        attributeMapping: {
+          givenName: ProviderAttribute.FACEBOOK_NAME,
+          address: ProviderAttribute.other('fb-address'),
+          custom: {
+            customAttr1: ProviderAttribute.FACEBOOK_EMAIL,
+            customAttr2: ProviderAttribute.other('fb-custom-attr'),
+          },
+        },
+      });
+
+      // THEN
+      expect(stack).toHaveResource('AWS::Cognito::UserPoolIdentityProvider', {
+        AttributeMapping: {
+          given_name: 'name',
+          address: 'fb-address',
+          customAttr1: 'email',
+          customAttr2: 'fb-custom-attr',
+        },
+      });
+    });
+  });
+
+  describe('google', () => {
+    test('defaults', () => {
+      // GIVEN
+      const stack = new Stack();
+      const pool = new UserPool(stack, 'userpool');
+
+      // WHEN
+      new UserPoolIdentityProviderGoogle(stack, 'userpoolidp', {
+        userPool: pool,
+        clientId: 'google-client-id',
+        clientSecret: 'google-client-secret',
+      });
+
+      expect(stack).toHaveResource('AWS::Cognito::UserPoolIdentityProvider', {
+        ProviderName: 'Google',
+        ProviderType: 'Google',
+        ProviderDetails: {
+          client_id: 'google-client-id',
+          client_secret: 'google-client-secret',
+          authorize_scopes: 'profile',
+        },
+      });
+    });
+
+    test('scopes', () => {
+      // GIVEN
+      const stack = new Stack();
+      const pool = new UserPool(stack, 'userpool');
+
+      // WHEN
+      new UserPoolIdentityProviderGoogle(stack, 'userpoolidp', {
+        userPool: pool,
+        clientId: 'google-client-id',
+        clientSecret: 'google-client-secret',
+        scopes: ['scope1', 'scope2'],
+      });
+
+      expect(stack).toHaveResource('AWS::Cognito::UserPoolIdentityProvider', {
+        ProviderName: 'Google',
+        ProviderType: 'Google',
+        ProviderDetails: {
+          client_id: 'google-client-id',
+          client_secret: 'google-client-secret',
+          authorize_scopes: 'scope1 scope2',
+        },
+      });
+    });
+
+    test('registered with user pool', () => {
+      // GIVEN
+      const stack = new Stack();
+      const pool = new UserPool(stack, 'userpool');
+
+      // WHEN
+      const provider = new UserPoolIdentityProviderGoogle(stack, 'userpoolidp', {
+        userPool: pool,
+        clientId: 'google-client-id',
+        clientSecret: 'google-client-secret',
+      });
+
+      // THEN
+      expect(pool.identityProviders).toContain(provider);
+    });
+
+    test('attribute mapping', () => {
+      // GIVEN
+      const stack = new Stack();
+      const pool = new UserPool(stack, 'userpool');
+
+      // WHEN
+      new UserPoolIdentityProviderGoogle(stack, 'userpoolidp', {
+        userPool: pool,
+        clientId: 'google-client-id',
+        clientSecret: 'google-client-secret',
+        attributeMapping: {
+          givenName: ProviderAttribute.GOOGLE_NAME,
+          address: ProviderAttribute.other('google-address'),
+          custom: {
+            customAttr1: ProviderAttribute.GOOGLE_EMAIL,
+            customAttr2: ProviderAttribute.other('google-custom-attr'),
+          },
+        },
+      });
+
+      // THEN
+      expect(stack).toHaveResource('AWS::Cognito::UserPoolIdentityProvider', {
+        AttributeMapping: {
+          given_name: 'name',
+          address: 'google-address',
+          customAttr1: 'email',
+          customAttr2: 'google-custom-attr',
+        },
+      });
+    });
+  });
+});
